+++ packages/generator-custom-element/Changelog.md
@@ -1,5 +1,11 @@
 # Yeoman generator - custom element
 
+## [0.0.2] - 2019-01-15
+### Added
+Generator is able to create a `unit test` file. 
+### Changed
+Unit test related with creating an element uses much complex reg expressions.
+
 ## [0.0.1] - 2019-01-14
 ### Added
 Generator is able to create a simple custom element v1